requirements.txt
altgraph==0.17.3
auto-py-to-exe==2.27.0
bottle==0.12.23
bottle-websocket==0.2.9
cffi==1.15.1
contourpy==1.0.7
cycler==0.11.0
Eel==0.14.0
fonttools==4.38.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
imageio==2.25.0
kiwisolver==1.4.4
matplotlib==3.6.3
networkx==3.0
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
pefile==2022.5.30
Pillow==9.4.0
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
pywin32-ctypes==0.2.0
scikit-image==0.19.3
scipy==1.10.0
six==1.16.0
tifffile==2023.1.23.1
whichcraft==0.6.1
zope.event==4.6
zope.interface==5.5.2
